+++ R/draggable_buckets.R
@@ -53,12 +53,10 @@ draggable_buckets <- function(input_id, label, elements = character(), buckets) 
   elements_iterator$it <- 0
 
   shiny::tagList(
-    shiny::tags$head(shiny::singleton(
+    shiny::singleton(shiny::tags$head(
       shiny::includeScript(system.file("widgets/draggable_buckets.js", package = "teal.widgets"))
     )),
-    shiny::tags$head(shiny::singleton(
-      shiny::includeCSS(system.file("css/draggable_buckets.css", package = "teal.widgets"))
-    )),
+    include_css_files("draggable_buckets.css"),
     shiny::div(
       shiny::tags$span(label),
       render_unbucketed_elements(elements = elements, elements_iterator = elements_iterator, widget_id = input_id),